- Refactor code structure - Added new packages: model, dao, repository, services, controller
