+++ backend/src/main/java/pl/roseitconsulting/roseitapp/model/Customer.java
@@ -1,4 +1,4 @@
-package pl.roseitconsulting.roseitapp.customer;
+package pl.roseitconsulting.roseitapp.model;
 
 import jakarta.persistence.*;
