+++ backend/src/main/java/pl/roseitconsulting/roseitapp/repository/CustomerRepository.java
@@ -1,7 +1,8 @@
-package pl.roseitconsulting.roseitapp.customer;
+package pl.roseitconsulting.roseitapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import pl.roseitconsulting.roseitapp.model.Customer;
 
 @Repository
 public interface CustomerRepository extends JpaRepository<Customer, Integer> {
