+++ backend/src/test/java/pl/roseitconsulting/roseitapp/customer/CustomerJPADataAccessServiceTest.java
@@ -1,14 +1,14 @@
 package pl.roseitconsulting.roseitapp.customer;
 
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
 import org.mockito.junit.jupiter.MockitoExtension;
+import pl.roseitconsulting.roseitapp.dao.CustomerJPADataAccessService;
+import pl.roseitconsulting.roseitapp.model.Customer;
+import pl.roseitconsulting.roseitapp.repository.CustomerRepository;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
