+++ backend/src/test/java/pl/roseitconsulting/roseitapp/customer/CustomerRepositoryTest.java
@@ -1,11 +1,13 @@
 package pl.roseitconsulting.roseitapp.customer;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import pl.roseitconsulting.roseitapp.AbstractTestcontainers;
+import pl.roseitconsulting.roseitapp.model.Customer;
+import pl.roseitconsulting.roseitapp.repository.CustomerRepository;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
