+++ backend/src/test/java/pl/roseitconsulting/roseitapp/customer/CustomerRowMapperTest.java
@@ -1,11 +1,12 @@
 package pl.roseitconsulting.roseitapp.customer;
 
-import org.checkerframework.checker.units.qual.C;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import pl.roseitconsulting.roseitapp.dao.CustomerRowMapper;
+import pl.roseitconsulting.roseitapp.model.Customer;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
