+++ backend/src/test/java/pl/roseitconsulting/roseitapp/customer/CustomerServiceTest.java
@@ -1,19 +1,22 @@
 package pl.roseitconsulting.roseitapp.customer;
 
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import pl.roseitconsulting.roseitapp.dao.CustomerDao;
 import pl.roseitconsulting.roseitapp.exception.DuplicateResourceException;
 import pl.roseitconsulting.roseitapp.exception.ResourceNotFoundException;
+import pl.roseitconsulting.roseitapp.model.Customer;
+import pl.roseitconsulting.roseitapp.model.CustomerRegistrationRequest;
+import pl.roseitconsulting.roseitapp.model.CustomerUpdateRequest;
+import pl.roseitconsulting.roseitapp.services.CustomerService;
 
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
